requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
python-dotenv==1.0.0
SQLAlchemy==2.0.21
Werkzeug==2.3.7
python-jose==3.3.0
cryptography==41.0.3
Flask-Limiter==3.5.0
bleach==6.0.0
Flask-Admin==1.6.1
Flask-Login==0.6.3
better-profanity==0.7.0
Flask-Bootstrap==3.3.7.1
